pyproject.toml
[project]
name = "gmail-top-sender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.156.0",
    "google-auth-oauthlib>=1.2.1",
    "ruff>=0.8.4",
]


[tool.ruff.lint]
# https://docs.astral.sh/ruff/settings/#select
# default for select is ["E4", "E7", "E9", "F"]
# we add D for docstrings, and N for naming conventions
extend-select = ["D", "N"]

[tool.ruff.lint.pydocstyle]
convention = "google"
